Refactored window extraction package
* also added extractor skelton
