+++ src/main/groovy/org/anc/lapps/chunk/window/extraction/SentBasedWindow.groovy
@@ -1,0 +1,7 @@
+package org.anc.lapps.chunk.window.extraction
+
+/**
+ * Created by krim on 8/16/2016.
+ */
+class SentBasedWindow implements WindowExtractionI {
+}
