+++ src/main/groovy/org/anc/lapps/chunk/window/extraction/WindowExtractionI.groovy
@@ -1,0 +1,7 @@
+package org.anc.lapps.chunk.window.extraction
+
+/**
+ * Created by krim on 8/17/2016.
+ */
+interface WindowExtractionI {
+}
